requirements.txt
requests==2.28.1
dataclasses-json==0.5.7
pycryptodomex==3.15.0
PyYAML==6.0
